+++ app/src/main/java/nextstep/shoppingcart/NavigationItem.kt
@@ -1,19 +1,10 @@
 package nextstep.shoppingcart
 
-import androidx.compose.foundation.layout.padding
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowBack
-import androidx.compose.material3.ExperimentalMaterial3Api
-import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
-import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Text
-import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
+import nextstep.shoppingcart.cart.CartScreen
 import nextstep.shoppingcart.products.ProductItemUiState
 import nextstep.shoppingcart.products.ProductsScreen
 import nextstep.shoppingcart.products.StubProductItemUiStates
@@ -57,30 +48,5 @@ fun AppNavHost(
         composable(NavigationItem.Cart.route) {
             CartScreen(navHostController)
         }
-    }
-}
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun CartScreen(navController: NavHostController) {
-    Scaffold(
-        topBar = {
-            TopAppBar(
-                title = { Text(text = "장바구니") },
-                navigationIcon = {
-                    IconButton(onClick = { navController.popBackStack() }) {
-                        Icon(
-                            imageVector = Icons.Filled.ArrowBack,
-                            contentDescription = "뒤로가기"
-                        )
-                    }
-                },
-            )
-        }
-    ) { innerPadding ->
-        Text(
-            text = "CartScreen",
-            modifier = Modifier.padding(innerPadding)
-        )
     }
 }
